+++ src/types/AttributeOrAttributeValueFilter.ts
@@ -1,4 +1,4 @@
-export type AttributeKey = string & {readonly ___attributeKey: unique symbol};
+export type AttributeKey = string;
 
 export type AttributeFilter = AttributeKey;
 export type AttributeWithValuesFilter =
@@ -13,6 +13,11 @@ export type AttributeWithValuesFilter =
       id: number;
       key?: undefined;
       values: Array<string | number>;
+    }
+  | {
+      type: 'attributes:not';
+      key: string;
+      values: Array<number>;
     };
 
 export interface AttributeWithBooleanValueFilter {