+++ src/types/Product.ts
@@ -184,7 +184,7 @@ export interface BapiPrice {
 }
 
 export interface AppliedReduction {
-  category: 'sale' | 'campaign' | 'voucher';
+  category: 'sale' | 'campaign' | 'promotion' | 'voucher';
   type: 'relative' | 'absolute';
   amount: {
     // 0-1